# make pattern_* and makeiplist tools faster