--- a/tools/recon/makeiplist.rb
+++ b/tools/recon/makeiplist.rb
@@ -24,7 +24,6 @@ end
 
 $:.unshift(File.expand_path(File.join(File.dirname(msfbase), '..', '..', 'lib')))
 
-require 'msfenv'
 require 'rex'
 require 'optparse'