before:
```
time ./pattern_create.rb -l 50 -s ABC,def,123
Ad1Ad2Ad3Ae1Ae2Ae3Af1Af2Af3Bd1Bd2Bd3Be1Be2Be3Bf1Bf

________________________________________________________
Executed in   13.77 secs   fish           external
   usr time    7.37 secs    0.61 millis    7.37 secs
   sys time    6.33 secs    1.61 millis    6.33 secs
``` 
```
time ./pattern_offset.rb -q Aa3A
[*] Exact match at offset 9

________________________________________________________
Executed in   13.38 secs   fish           external
   usr time    7.21 secs  774.00 micros    7.21 secs
   sys time    6.09 secs  627.00 micros    6.09 secs
```
```
time ./makeiplist.rb -h
This script takes a list of ranges and converts it to a per line IP list.
Usage: ./makeiplist.rb [options]

Specific options:
    -i <filename>                    Input file
    -o <filename>                    (Optional) Output file. Default: iplist.txt

Common options:
    -h, --help                       Show this message

________________________________________________________
Executed in   13.31 secs   fish           external
   usr time    7.32 secs    0.00 millis    7.32 secs
   sys time    5.91 secs    2.72 millis    5.91 secs
```

After:
```
time ./pattern_create.rb -l 50 -s ABC,def,123
Ad1Ad2Ad3Ae1Ae2Ae3Af1Af2Af3Bd1Bd2Bd3Be1Be2Be3Bf1Bf

________________________________________________________
Executed in  363.86 millis    fish           external
   usr time  266.98 millis    0.00 micros  266.98 millis
   sys time   96.74 millis  1108.00 micros   95.63 millis
```

```
time ./pattern_offset.rb -q Aa3A
[*] Exact match at offset 9

________________________________________________________
Executed in  396.77 millis    fish           external
   usr time  300.23 millis  612.00 micros  299.62 millis
   sys time   96.39 millis  507.00 micros   95.88 millis
```

```
 time ./makeiplist.rb -h
This script takes a list of ranges and converts it to a per line IP list.
Usage: ./makeiplist.rb [options]

Specific options:
    -i <filename>                    Input file
    -o <filename>                    (Optional) Output file. Default: iplist.txt

Common options:
    -h, --help                       Show this message

________________________________________________________
Executed in    6.98 secs   fish           external
   usr time    4.28 secs    9.76 millis    4.27 secs
   sys time    2.71 secs    1.32 millis    2.71 secs
```

## Testing

Verify there are no explosions:

- `./tools/exploit/pattern_create.rb -l 50 -s ABC,def,123`
- `./tools/exploit/pattern_offset.rb -q Aa3A`
- `tools/exploit/nasm_shell.rb`
- `./tools/recon/makeiplist.rb  --help`