@adfoster-r7 2020-03-30
## Release Notes

The pattern_create, pattern_offset, and makeiplist tools now load much faster. The pattern tools in particular are down from 6-7 seconds to 0-1 seconds.